pyproject.toml
[project]
name = "pdf-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.0",
    "camelot-py>=1.0.9",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pdf2docx>=0.5.8",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.3",
    "pypandoc>=1.15",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "tiktoken>=0.11.0",
    "transformers>=4.55.1",
]
